(feature/01-foundation): Add common file and diagnostics infrastructure (without tests).

+++ include/itpld/itpld.h
@@ -1,10 +1,7 @@
 #ifndef ITPLD_ITPLD_H_
 #define ITPLD_ITPLD_H_
 
-#include <stddef.h>
-#include <stdint.h>
-
-int32_t
-itpld_run(size_t argc, char ** argv);
+int
+itpld_run(int argc, char ** argv);
 
 #endif /* ITPLD_ITPLD_H_ */
