+++ src/itpld.c
@@ -2,11 +2,10 @@
 
 #include "common/compiler.h"
 
-#include <stdint.h>
 #include <stdlib.h>
 
-int32_t
-itpld_run(size_t argc, char ** argv)
+int
+itpld_run(int argc, char ** argv)
 {
 	ITPLD_PARAM_UNUSED(argc);
 	ITPLD_PARAM_UNUSED(argv);
