+++ src/main.c
@@ -1,9 +1,7 @@
 #include "itpld/itpld.h"
 
-#include <stddef.h>
-
 int
 main(int argc, char ** argv)
 {
-	return (int)itpld_run((size_t)argc, argv);
+	return itpld_run(argc, argv);
 }
